+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ CHANGELOG
 * feature: Add APIs to export Airflow training, tuning and model config
 * doc-fix: Fix typos in tensorflow serving documentation
 * doc-fix: Add estimator base classes to API docs
+* feature: HyperparameterTuner: add support for Automatic Model Tuning's Warm Start Jobs
 
 1.14.2
 ======